minesweeper_test.py
import pytest
from minesweeper import Minesweeper

def test_create_minefield():
    minesweeper = Minesweeper(5, 3)
    minesweeper.create_minefield()
    # Add assertions based on the expected output of create_minefield

def test_set_bombs():
    minesweeper = Minesweeper(5, 3)
    minesweeper.set_bombs()
    # Add assertions based on the expected state of the minesweeper instance

def test_set_values():
    minesweeper = Minesweeper(5, 3)
    minesweeper.set_values()
    # Add assertions based on the expected state of the minesweeper instance

def test_nbors():
    minesweeper = Minesweeper(5, 3)
    minesweeper.vis = []  # Reset vis
    minesweeper.nbors(1, 1)
    # Add assertions based on the expected state of the minesweeper instance after calling nbors

def test_check_over():
    minesweeper = Minesweeper(5, 3)
    minesweeper.values_of_bombs = [['M', ' ', ' ', ' ', ' '],
                                   [' ', ' ', ' ', ' ', ' '],
                                   [' ', ' ', ' ', ' ', ' '],
                                   [' ', ' ', ' ', ' ', ' '],
                                   [' ', ' ', ' ', ' ', ' ']]
    assert minesweeper.check_over() == False
    # Add more assertions based on different scenarios

def test_show_bombs():
    minesweeper = Minesweeper(5, 3)
    minesweeper.show_bombs()
    # Add assertions based on the expected state of the minesweeper instance

def test_play_game():
    # This test is more complex and may involve mocking user input to simulate a game
    pass